fix: ajustando campos listagem de pets
